+++ games/maps/js/map/map-init.js
@@ -1,4 +1,4 @@
-import { buildBaseStyle } from "./map-styles.js?v=2026-08-26.27";
+import { buildBaseStyle } from "./map-styles.js?v=2026-08-26.28";
 
 export function createMap(initialSettings) {
   const map = new maplibregl.Map({